==> notebooks/Spiketrain Props/01 - Spiketrain Stats.ipynb <==
from typing import List
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import seaborn as sns
from drn_interactions.transforms import SpikesHandler
from drn_interactions.spiketrains.spiketrain_stats import SpikeTrainDescriptor
from IPython.display import display
from drn_interactions.config import Config

sns.set_theme(context="talk", style="ticks")

%load_ext autoreload
%autoreload 2

sh = SpikesHandler(block="pre", bin_width=1, t_start=0, t_stop=1800)

df_spikes = sh.spikes

descriptor = SpikeTrainDescriptor()

df_stats = descriptor.describe(
    df_spikes, 
    t_start=0, 
    t_stop=1800, 
    burst_thresh=0.02, 
    max_fr_binsize=30, 
    cv_fr_binsize=30, 
    mfr_bin_size=30, 
    mfr_exclude_below=0.5,
    )

display(df_stats.sample(5))


df_stats.to_csv(Config.derived_data_dir / "spiketrain_stats.csv", index=False)

from binit import which_bin
from drn_interactions.spiketrains.spiketrain_stats import SpikeTrainStats




from drn_interactions.spiketrains.spiketrain_stats import cv2

bins = np.arange(0, 1801, 30)
df_spikes["segment"] = which_bin(df_spikes["spiketimes"].values, bins, time_after=30)
stats = SpikeTrainStats(thresh_burst=0.02)
df_res = df_spikes.groupby(["neuron_id", "segment"])["spiketimes"].apply(stats)
df_res2 = df_res.reset_index().rename(columns={"level_2": "metric"}).assign(spiketimes=lambda x: x["spiketimes"].astype(float))

df_medians =df_res2.groupby(["neuron_id", "metric"])["spiketimes"].apply(lambda x: np.nanmedian(x)).reset_index()
df_medians = df_medians.pivot(index="neuron_id", columns="metric", values="spiketimes")
df_medians.reset_index().to_csv(Config.derived_data_dir / "spiketrain_stats_segments.csv", index=False)


df_medians =df_res2.groupby(["neuron_id", "metric"])["spiketimes"].apply(lambda x: np.quantile(x, 0.25)).reset_index()
df_medians = df_medians.pivot(index="neuron_id", columns="metric", values="spiketimes")
df_medians.reset_index().to_csv(Config.derived_data_dir / "spiketrain_stats_segments_lower.csv", index=False)

from scipy.stats import variation
df_cvs =df_res2.fillna(method="backfill").fillna(method="bfill").groupby(["neuron_id", "metric"])["spiketimes"].apply(lambda x: variation(x)).reset_index()
df_cvs = df_cvs.pivot(index="neuron_id", columns="metric", values="spiketimes")
df_cvs.reset_index().to_csv(Config.derived_data_dir / "spiketrain_stats_volitility.csv", index=False)

df_cvs
