==> notebooks/Spiketrain Props/04 - Neuron Props Numbers.ipynb <==
from drn_interactions.io import load_recordings, load_neurons, load_derived_generic
from drn_interactions.stats import se_mean, se_prop, p_adjust, prop_of_total
from drn_interactions.spiketrains.neurontype_props import AnovaPostHoc, ChiSquarePostHoc
from IPython.display import display
import numpy as np
from scipy.stats import chi2_contingency
import pandas as pd
import pingouin as pg
from itertools import combinations

%load_ext autoreload
%autoreload 2

def mean_se(x, round=1):
    m = np.mean(x)
    se = se_mean(x)
    return f"{m:2.{round}f} ± {se:2.{round}f}"

def mean_se_prop(x, round=2):
    p = np.mean(x) * 100
    se = se_prop(x) * 100
    return f"{p:2.{round}f} ± {se:2.{round}f}"


def compare_across_neuron_types(df, col, round=2):
    """
    Compare the mean and SE of a column across clusters.
    """
    df = df.copy()
    df["cluster"] = df["cluster"].astype(str)
    df = df.groupby(["cluster", col]).mean().reset_index()
    df = df.pivot(index="cluster", columns=col, values="mean")
    df.columns = [f"{col} {c}" for c in df.columns]
    df["se"] = df.apply(lambda x: se_mean(x), axis=1)
    df["prop"] = df.apply(lambda x: np.mean(x) * 100, axis=1)
    df["se_prop"] = df.apply(lambda x: se_prop(x) * 100, axis=1)
    df = df.round(round)
    return df


# number of recordings

df_recordings = load_recordings()

print(f"Total Recprdings: {len(df_recordings)}")


num_by_proto = df_recordings.groupby("experiment_name").apply(len).to_frame("Number of Recordings")
num_by_proto.index.name = "Recording Protocol"
display(num_by_proto)

# number of neurons per recordings and overall

neurons = load_neurons()

total_neurons = len(neurons)
print(f"Total Neurons: {total_neurons}")

num_by_proto = neurons.groupby("experiment_name").apply(len).to_frame("Number of Neurons")
num_by_proto.index.name = "Recording Protocol"
display(num_by_proto)

num_by_session = neurons.groupby("session_name").apply(len)
print(f"Number of neurons per session: {mean_se(num_by_session)}")

# proportion of units of each type

clusters = load_derived_generic("neuron_types.csv")
clusters = clusters.merge(neurons[["neuron_id", "session_name"]])

mean_yeild_per_session = (
    clusters.groupby(["session_name", "neuron_type"]).apply(len)
    .to_frame("N").reset_index()
    .groupby("neuron_type")["N"].apply(mean_se)
    .to_frame("Mean Yeild")
)
mean_yeild_per_session.index.name = "Neuron Type"
display(mean_yeild_per_session)


prop_total_by_session = (
    clusters
    .pivot_table(index="session_name", columns="neuron_type", values="neuron_id", aggfunc=len)
    .fillna(0)
    .transform(prop_of_total, axis=1)
    .reset_index()
    .melt(id_vars="session_name", var_name="neuron_type", value_name="Proportion of Total")
)
prop_total_by_session.groupby("neuron_type", as_index=False)["Proportion of Total"].apply(mean_se_prop)





df_single_unit = load_derived_generic("neuron_types_single_unit.csv").assign(**{"Electrode Type": "Single Unit"})
df_probe = load_derived_generic("neuron_types.csv").assign(**{"Electrode Type": "Silicon Probe"})

df = (
    pd.concat([df_single_unit[["Electrode Type", "neuron_type"]], df_probe[["Electrode Type", "neuron_type"]]])
    .pipe(lambda x: pd.crosstab(x["neuron_type"], x["Electrode Type"]))
)
stat, p, dof, expected = chi2_contingency(df)

props = df.apply(prop_of_total).round(2)
display(props)
star = "*" if p < 0.05 else ""
print(f"Single Unit v Probe Neuron Type, Chi Square Test: \n\tChi-Square({dof}) = {stat:.1f}\n\tp = {p:.2f}{star}")



clusters = load_derived_generic("neuron_types.csv").merge(load_neurons()[["neuron_id", "session_name"]])
spikes = load_derived_generic("spiketrain_stats_segments.csv")
waveforms = load_derived_generic("waveform_summary.csv")

meta_cols = ["neuron_id", "session_name", "neuron_type"]
spiketrain_stats = (
    spikes
    .merge(clusters)
    .merge(waveforms, how="outer")
    .melt(id_vars=meta_cols, var_name="metric", value_name="value")
    .dropna()
    .assign(value=lambda x: x.value.astype(float))
)

# numeric metrics
anova_calculator = AnovaPostHoc(neuron_type_col="neuron_type", value_col="value", round=1)
numeric_metrics = ["cv_isi", "cv_isi_burst", "mean_firing_rate", "median_burst_interval", "width_minpost",]
df_res_numeric = (
    spiketrain_stats
    .query(f"metric in @numeric_metrics")
    .groupby("metric")
    .apply(anova_calculator)
)

# categorical metrics
prop_calculator = ChiSquarePostHoc(neuron_type_col="neuron_type", value_col="value", round=1)
categorical_metrics = ["is_burst", "has_basepre", "has_basepost"]
df_res_categorical = (
    spiketrain_stats
    .query(f"metric in @categorical_metrics and value in (0, 1)")
    .groupby("metric")
    .apply(prop_calculator)
)

# display results
display(df_res_numeric)
display(df_res_categorical)

# volitility of spike train properties by neuron type

spikes_vol = load_derived_generic("spiketrain_stats_volitility.csv")

meta_cols = ["neuron_id", "session_name", "neuron_type"]
spiketrain_vol = (
    spikes_vol
    .merge(clusters)
    .melt(id_vars=meta_cols, var_name="metric", value_name="value")
    .assign(value=lambda x: x.value.astype(float))
)

# numeric metrics
anova_calculator = AnovaPostHoc(neuron_type_col="neuron_type", value_col="value", round=1)
numeric_metrics = ["cv_isi", "cv_isi_burst", "mean_firing_rate", ]
df_res_numeric = (
    spiketrain_vol
    .query(f"metric in @numeric_metrics")
    .groupby("metric")
    .apply(anova_calculator)
)

# categorical metrics
prop_calculator = ChiSquarePostHoc(neuron_type_col="neuron_type", value_col="value", round=1)
categorical_metrics = ["is_burst"]
df_res_categorical = (
    spiketrain_vol
    .query(f"metric in @categorical_metrics and value in (0, 1)")
    .groupby("metric")
    .apply(prop_calculator)
)

# display results
display(df_res_numeric)
display(df_res_categorical)
